refactor(trigger): extract daemon lifecycle into daemon.go
Pure move of the standalone daemon (always-on task) lifecycle out of
engine.go: registerDaemon, startDaemon, onDaemonRunFinished (exit
handling + restart policy), and the restart backoff constants.
Crash-loop detection stays in crashloop.go and WebUI state tracking in
daemon_state.go, respecting the existing boundaries. No behavior
change; code relocated verbatim.

Part of #388 (god-file decomposition).
